Add more buffer to the embedding llm's context window
Something is still going on with the tokenizer for the embeddings model.
It's underestimating tokens for the prompt and file input.

This commit *works around* the issue by lowering the number of tokens
being sent per batch to the model, but it doesn't get at the root cause
of the need to do so.