Implement fractional backoff in AI.Accumulator
We estimate token size. When it gets close, it's likely to be inaccurate
enought to result in errors on large inputs. This change implements a
retry in AI.Accumulator, backing off from the original context window
size by 20% for 3 attempts, trying to ease into a chunk size that works
for the input.